Fixes optional steps for github workflow

+++ growlithe/graph/parsers/sam.py
@@ -459,8 +459,10 @@ class SAMParser:
             self.config.growlithe_path, "layers", "pydatalog.zip"
         )
         os.makedirs(destination, exist_ok=True)
-        shutil.copy(self.config.pydatalog_layer_path, destination)
-
+        if os.exists(self.config.pydatalog_layer_path):
+            shutil.copy(self.config.pydatalog_layer_path, destination)
+        else:
+            logger.warning(f"No zip found at {self.config.pydatalog_layer_path}")
     def save_config(self):
         """
         Save the updated configuration to a YAML file in the growlithe folder.
