+++ growlithe/graph/parsers/terraform.py
@@ -278,7 +278,10 @@ class TerraformParser:
             self.config.growlithe_path, "layers", "pydatalog.zip"
         )
         os.makedirs(destination, exist_ok=True)
-        shutil.copy(self.config.pydatalog_layer_path, destination)
+        if os.exists(self.config.pydatalog_layer_path):
+            shutil.copy(self.config.pydatalog_layer_path, destination)
+        else:
+            logger.warning(f"No zip found at {self.config.pydatalog_layer_path}")
 
     def save_config(self):
         """
